final Commit Added buttons Added counter for  number of each items
Telling to run the setupCart function when page is fully loaded. Defined the event handlers for the add to order buttons. Got all content from the html document that has the classname of addButton. Added items to the shopping cart. Added buttons and Added a counter for the number of each items that are in the shopping cart.  Added comments. Ready for review.